Refactor GridRouter class and add necessary methods
This commit refactors the GridRouter class by adding necessary methods for grid routing. The bfs method is implemented to find the shortest path between two terminals using a breadth-first search algorithm. The reconstructPath method is added to reconstruct the path from end to start. The isValid method is created to check if a move is valid within the grid bounds and not a barrier. The markPath method is implemented to mark the path in the grid and calculate its length. The findTerminals method is added to find terminals by color. The solveGrid method is created as the main function to run the routing algorithm. The outputResult method is implemented to output the result to a file. The main method is updated to read the input grid from a file and solve the grid routing problem.